+++ src/main/java/com/quasiris/qsf/query/SearchQueryMapper.java
@@ -21,7 +21,9 @@ public class SearchQueryMapper {
         searchQuery.setTracking(searchQueryDTO.getTracking());
         searchQuery.setRequestOrigin(searchQueryDTO.getRequestOrigin());
         searchQuery.setRows(searchQueryDTO.getRows());
-        searchQuery.setTrackingTags(new ArrayList<>(searchQueryDTO.getTrackingTags()));
+        if(searchQueryDTO.getTrackingTags() != null) {
+            searchQuery.setTrackingTags(new ArrayList<>(searchQueryDTO.getTrackingTags()));
+        }
 
         Sort sort = map(searchQueryDTO.getSort());
         searchQuery.setSort(sort);